Remove bug
When called more times than it can launch  program (not enough command in rc
file) the program was not displaying a message first, only an blank line.
This was found thanks to @RomFouq.

Number was already incremented. I add a more strict condition to solve the
problem.

TODO: Use more explicit code.