Use more laziness in constants to limit errors.
Related to #3

It should limit error displayed when it is not necessary.
For example, when using command to list items, we don't need editor and that
shouldn't print an error message. However, when using a command to modify an entry, we
need an editor and we display the error message.